Overwrite sizeHint() for the editor to handle missing window state information.
This triggers only in case the configuration file has no window state
information (or no configuration file at all).
When this happens, the editor would default to a very ugly width due to
the dock widget layout. This overwrites the value reported via sizeHint()
to a width a bit smaller than half the main window size (either the one
loaded from the configuration or the default value of 800).
The height is only a dummy value which will be essentially ignored by
the layouting as the editor is set to expand to fill the available space.